split output test into unit and integration
